refactor(e2e): categorize baseline smoke tests into subdirectories
Split the flat smoke/ dir by concern: pull the static/infra harness
self-tests (registry guard, provisioning, user_ops) out into a sibling
guards/ dir, and group the remaining live worked-examples under smoke/
by subject — samples/ (shared helpers), matrix/, behavior/, inspection/,
adapters/. Repoint helper imports to smoke.samples, fix self-referential
run-with docstrings, and update README + ADDING_AN_ADAPTER paths.

No test bodies, fixtures, or toolkit modules change; CI needs no edit
(it runs pytest tests/e2e/baseline/ recursively). Collection unchanged
at 79 tests.
