+++ tests/e2e/baseline/guards/test_adapter_registry.py
