+++ tests/e2e/baseline/guards/test_user_ops.py
