+++ tests/e2e/baseline/smoke/behavior/test_agent_scenarios.py
