+++ tests/e2e/baseline/smoke/inspection/test_events.py
@@ -21,7 +21,7 @@ import pytest
 from band.core.types import MessageType
 
 from tests.e2e.baseline.agents import Adapter, with_agents
-from tests.e2e.baseline.smoke.sample_agents import (
+from tests.e2e.baseline.smoke.samples.sample_agents import (
     TOOL_AGENT,
     emit_event_instruction,
     emit_thoughts_instruction,
